Implements review suggestion
[Desc]
 - Creating a new NumberFormat instance on
every call to formatAsCurrencyBr() is inefficient,
as NumberFormat.getCurrencyInstance() can be a
relatively expensive operation. Since this function
is used in Composables, it can be called frequently
during recompositions. It's better to cache the
NumberFormat instance at the file level using lazy initialization.
 - item.shoppList.budget.formatAsCurrencyBr() is called multiple
times (in the if condition and inside the if/else blocks).
It should be called once and its result stored in a
variable to avoid re-computation.